refactor: move buildKatasCollectionSchema from data/katas to lib/schema-org
SEO/JSON-LD generation is presentation logic, not data. Extracted to
lib/schema-org.ts alongside a dedicated test file move. data/katas.ts
now contains only the data and sortedKatas helper.
