+++ src/pages/en/katas/index.astro
@@ -1,7 +1,7 @@
 ---
 import Layout from '../../../layouts/Layout.astro';
 import KatasPage from '../../../components/katas/KatasPage.astro';
-import { buildKatasCollectionSchema } from '../../../data/katas';
+import { buildKatasCollectionSchema } from '../../../lib/schema-org';
 import { t } from '../../../i18n/utils';
 
 const lang = 'en';
